Make temporary directory for cleaned code
After generating code user can clean it up using the script which is
edited with this commit. The cleaned code will be put inside
"code_gen_temp" directory. User can use then 'meld' or other diff-tool
to apply changes in a neat way to the main code. 'dos2unix', 'vim' must
be installed on the host.