New approach in counting number of received bits
Earlier there was an interrupt which was invoked with frequency equal to
UART bit duration. In each interrupt a variable was incremented to count
how many bits were received between two next slopes. This is redundant
as one can simply check the counter value for the RX timer and divide
its value by the value of counter after one bit duration. There are
margins used because it is obvious that the value of the counter may
vary at each period, because it depends on the environment. App gets
simpler thanks to that change because we use one interrupt less and one
header file could be deleted.