Fix errors with TX when often transmitting
When data is being transmitted and transmit_data is invoked then there
was a bug which causes transmitting to fail. This is caused by resetting
of the counter value when invoking transmit_data. Now everything works
well because it is checked inside the function which enables TX ISR
whether the interrupt is enabled. If it is enabled function does
nothing.